video-physics-v0/requirements.txt
opencv-python
numpy
pandas
scipy
matplotlib
pysindy
pyyaml